Split up blocks and model
